+++ met4j-toolbox/src/main/java/fr/inrae/toulouse/metexplore/met4j_toolbox/GenerateDoc.java
@@ -64,26 +64,152 @@ public class GenerateDoc {
 
     private static String install="" +
             "```\n" +
+            "git clone https://forgemia.inra.fr/metexplore/met4j.git;\n" +
+            "cd met4j;\n" +
+            "mvn clean install \n" +
+            "\n" +
             "cd met4j-toolbox\n" +
             "mvn clean package\n" +
-            "```\n";
+            "```\n" +
+            "\n" +
+            "## Download executable jar from gitlab registry\n" +
+            "\n" +
+            "The executable jar is downloadable in the [met4j gitlab registry](https://forgemia.inra.fr/metexplore/met4j/-/packages).\n\n";
 
     private static String usage="" +
             "The toolbox can be launched using\n" +
             "```\n" +
-            "java -jar met4j-toolbox-<version>.jar\n" +
+            "java -jar target/met4j-toolbox-<version>.jar\n" +
             "```\n" +
             "which will list all the contained applications that can be called using\n" +
             "\n" +
             "```\n" +
-            "java -cp met4j-toolbox-<version>.jar <Package>.<App name> -h\n" +
-            "```\n";
+            "java -cp target/met4j-toolbox-<version>.jar <Package>.<App name> -h\n" +
+            "```\n" +
+            "\n" +
+            "Log4j from jsbml can be very verbose. You can make it silent by adding this command :\n" +
+            "\n" +
+            "```console\n" +
+            "java -Dlog4j.configuration= -cp target/met4j-toolbox-<version>.jar ...\n" +
+            "```\n" +
+            "\n" +
+            "## From singularity\n" +
+            "\n" +
+            "You need at least [singularity](https://sylabs.io/guides/3.5/user-guide/quick_start.html) v3.5.\n" +
+            "\n" +
+            "```console\n" +
+            "singularity pull met4j-toolbox.sif oras://registry.forgemia.inra.fr/metexplore/met4j/met4j-singularity:latest\n" +
+            "```\n" +
+            "\n" +
+            "If you want a specific version:\n" +
+            "\n" +
+            "```console\n" +
+            "singularity pull met4j-toolbox.sif oras://registry.forgemia.inra.fr/metexplore/met4j/met4j-singularity:x.y.z\n" +
+            "```\n" +
+            "\n" +
+            "If you want the last develop version:\n" +
+            "\n" +
+            "```console\n" +
+            "singularity pull met4j-toolbox.sif oras://registry.forgemia.inra.fr/metexplore/met4j/met4j-singularity:develop\n" +
+            "```\n" +
+            "\n" +
+            "If you want to build by yourself the singularity image:\n" +
+            "\n" +
+            "```console\n" +
+            "cd met4j-toolbox\n" +
+            "mvn package\n" +
+            "cd ../\n" +
+            "singularity build met4j-toolbox.sif met4j.singularity\n" +
+            "```\n" +
+            "\n" +
+            "\n" +
+            "This will download a singularity container met4j-toolbox.sif that you can directly launch.\n" +
+            "\n" +
+            "To list all the apps.\n" +
+            "```console\n" +
+            "met4j-toolbox.sif \n" +
+            "```\n" +
+            "\n" +
+            "To launch a specific app, prefix its name with the last component of its package name. For instance:\n" +
+            "\n" +
+            "```console\n" +
+            "met4j-toolbox.sif convert.Tab2Sbml -h -in fic.tsv -sbml fic.sbml\n" +
+            "```\n" +
+            "\n" +
+            "By default, singularity does not see the directories that are not descendants of your home directory. To get the directories outside your home directory, you have to specify the SINGULARITY_BIND environment variable.\n" +
+            "At least, to get the data in the default reference directory, you have to specify:\n" +
+            "In bash:\n" +
+            "```console\n" +
+            "export SINGULARITY_BIND=/db\n" +
+            "```\n" +
+            "In csh or in tcsh\n" +
+            "```console\n" +
+            "setenv SINGULARITY_BIND /db\n" +
+            "```\n" +
+            "\n" +
+            "## From docker\n" +
+            "\n" +
+            "First install [Docker](https://www.docker.com/).\n" +
+            "\n" +
+            "Pull the latest met4j image:\n" +
+            "\n" +
+            "```console\n" +
+            "sudo docker pull metexplore/met4j:latest\n" +
+            "```\n" +
+            "\n" +
+            "If you want a specific version:\n" +
+            "\n" +
+            "```console\n" +
+            "sudo docker pull metexplore/met4j:x.y.z\n" +
+            "```\n" +
+            "\n" +
+            "If you want the develop version:\n" +
+            "```console\n" +
+            "sudo docker pull metexplore/met4j:develop\n" +
+            "```\n" +
+            "\n" +
+            "If you want to build by yourself the docker image:\n" +
+            "\n" +
+            "```console\n" +
+            "cd met4j-toolbox\n" +
+            "mvn package\n" +
+            "cd ../\n" +
+            "sudo docker build -t metexplore/met4j:myversion .\n" +
+            "```\n" +
+            "\n" +
+            "\n" +
+            "To list all the apps:\n" +
+            "```console\n" +
+            "sudo docker run metexplore/met4j:latest met4j.sh\n" +
+            "```\n" +
+            "\n" +
+            "Don't forget to map volumes when you want to process local files.\n" +
+            "Example:\n" +
+            "\n" +
+            "```console\n" +
+            "sudo docker run -v /home/lcottret/work:/work \\\n" +
+            " metexplore/met4j:latest met4j.sh convert.Sbml2Tab \\\n" +
+            " -in /work/toy_model.xml -out /work/toy_model.tsv\n" +
+            "```\n" +
+            "\n" +
+            "If you change the working directory, you have to specify \"sh /usr/bin/met4j.sh\":\n" +
+            "\n" +
+            "```console\n" +
+            "sudo docker run -w /work -v /home/lcottret/work:/work \\\n" +
+            " metexplore/met4j:latest sh /usr/bin/met4j.sh convert.Sbml2Tab \\\n" +
+            " -in toy_model.xml -out toy_model.tsv\n" +
+            "```\n" +
+            "\n" +
+            "### Galaxy instance\n" +
+            "\n" +
+            "[Galaxy](https://galaxyproject.org/) wrappers for met4j-toolbox apps are available in the [Galaxy toolshed](https://toolshed.g2.bx.psu.edu/) (master version) and in the [Galaxy test toolsdhed](https://testtoolshed.g2.bx.psu.edu/) (develop version).\n" +
+            "Wrappers launch the met4j singularity container, so the server where your Galaxy instance is hosted must have Singularity installed.\n";
 
     private static StringBuffer getHeader(){
         StringBuffer sb = new StringBuffer();
         sb.append("# "+mainTitle); sb.append("\n");
         sb.append(shortDesc); sb.append("\n");
-        sb.append("\n## Installation\n");
+        sb.append("\n## Installation from source\n");
         sb.append(install);
         sb.append("\n## Usage\n");
         sb.append(usage);